requirements.txt
playwright
openai
beautifulsoup4
lxml
python-dotenv
pytest
pytest-asyncio
pydantic
flask
